if moistureReading is good for 5 times. Sleep for 1 Day
So rather than check to see if the moisture is at an adequate level every second. It checks to see if it is stable 5 times within the set amount of time. It is then put to sleep for 1 Day if it the moisture is stable. After 1 day it wakes up and the process repeats. If the moisture = dry, the water pump will activatate.